requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
colorama==0.4.6
fastapi==0.135.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.2
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
psycopg2-binary==2.9.9
jinja2
sqlalchemy==2.0.23
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
asyncpg==0.29.0